pyproject.toml
[tool.poetry]
name = "mini-cpm-v-playground"
version = "0.1.0"
description = ""
authors = ["Cheese <11363971+cheese-git@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gradio = "^4.26.0"
torch = "^2.2.2"
transformers = "^4.39.3"
accelerate = "^0.29.2"
timm = "^0.9.16"
torchvision = "^0.17.2"
python-dotenv = "^1.0.1"
opencv-python-headless = "^4.9.0.80"
sentencepiece = "^0.2.0"
protobuf = "^5.26.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
